refactor: ugly move all talkie-specific code into a plugin
This separates the game-level stuff like setting up the camera and
"configure the timestep" from the talkie stuff.

Things are still a mess, but the main.rs is looking more svelte at least.
